@travisdowns 2025-10-04
> "2.3e+09" carries too little information. I'm not a fan of 352352984723945871259.23154123 precision-without-accuracy, but this is too little. Often changes in the 0.5% range are measurable, and this hides it.

Fair. In part this is to discourage you from writing tests which fail to normalize the result down to something reasonable which doesn't need this in the first place, where you'd get 6 sig figs. Right now the width is set to 7, which means with scientific notification there is only room for 2 sig figs,  since `.e+00` take the other 5. I can set this to 8 or 9 which give 3 or 4 sig figs. 3 is usually enough to measure a 0.5% change, but not a 0.1% one. WDYT?

Other option is to use a suffix like `K`, `M`, `G` to represent magnitudes, but this has the disadvantage of jumping by 3 digits every time, which eats into the 4 chars you save by not using scientific notation.